Splitted filter into multiple modules
Filters in testrun and reports listing have much in common but some
parts are also very different. Thus, it makes sense to have to separate
modules for generating these filters that inherit common code from a
third one. This is what we have now. To make the complexity of which
functions are callable easier to handle we now have a dispatch attribute
that contains links from filter keys to the associated handler
function.