[FEAT] Page titles
Added page titles for the most important pages. This fixes BZ #825. This
does not add a page title for every page but I think some rarely used
pages like testrun/idlist can live with the default title "Tapper". It
should be easy to expand for every page that still needs a title.