[FIX] unify navi for testruns and reports
The navi for testruns and reports is generated in the controller. In
both cases the template code iterates through a list of hashes. The
meaning of the keys in both hashes is identical. Thus, it makes sense to
have the same template for both navis and in fact it works well.