Issue Summary covers all reports of group
So far we only listed those not_ok lines from the
current report and you had to click into every
other one to get an overall impression of the
error scenario.

Now we show all issues of all reports of the current
group.

However, it makes the details page much longer
in case of error, not sure that is what we want,
but I don't want to throw more Javascript at it
before getting feedback.