[FIX] deflate all DateTime columns
Until now, all DateTime columns used the
default stringification which just luckily
matched the string format of the used DB
systems (mostly mysql).

Triggered by a change in DBIx::Class now
this missing explicit deflate fails in
some subsequent Data::Compare function.

Because it was never expected to work it's
now on Tapper's side to fix that.

For the deflate we have to use the date-formatter
of the actually used DB driver and therefore had
to extend some functions to make the current
context $c available which gets us to the
used store and it's driver.